Revert "#16 feat: Adicionada opção para alterar status da O.S."
This reverts commit fe2bf77aa6d0d3b28e80c1f742da8dd7026f6679.
